Revert "refactor module"
