+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "ytml"
-version = "0.2.0"
+version = "0.1.3"
 edition = "2021"
 description = "A markup language that offers a different syntax to html"
 license = "MIT"
@@ -10,15 +10,15 @@ file-handling = []
 
 [[example]]
 name = "read_file"
-required-features = ["fs"]
+required-features = ["file-handling"]
 
 [[example]]
 name = "using_file_paths"
-required-features = ["fs"]
+required-features = ["file-handling"]
 
 [[example]]
 name = "write_html"
-required-features = ["fs"]
+required-features = ["file-handling"]
 
 [dependencies]
 pest = "2.5.5"
